+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Added:
 - Show channel mode in panel if available
 - Configuration options for font weight and bold weight
 - `/delay <seconds>` command that can be used to control the timing of commands sent on connect
+- Configuration option for the ordering of servers in the sidebar
 
 Fixed: